feat: leverages useContext to separate concerns
- moves all global state and functionality related to collections into a
  context provider
